+++ app/interactors/audio_script_processor/handle_audio_button.rb
@@ -5,7 +5,7 @@ module AudioScriptProcessor
 
     AUDIO_BUTTON = Audio::VoiceCatalog::DEFAULT_VOICE_SLUG.to_s.freeze
 
-    delegate :chat_id, :prompt_message, to: :context
+    delegate :chat_id, :prompt_message, :voice, to: :context
 
     def call
       context.fail!(error: ParentNotFoundError) if bot_message.blank?
@@ -17,7 +17,7 @@ module AudioScriptProcessor
 
     memoize def handled_button
       MediaGenerator::ButtonHandler::HandleButton.call(
-        button_request: AUDIO_BUTTON,
+        button_request: audio_button,
         chat_id:,
         tg_message_id: bot_message.tg_message_id,
         callback_query_id: nil
@@ -26,6 +26,10 @@ module AudioScriptProcessor
 
     def bot_message
       prompt_message.reload.bot_telegram_message
+    end
+
+    def audio_button
+      voice.presence || AUDIO_BUTTON
     end
   end
 end
