+++ app/jobs/script_generator/for_motivation/generate_motivation_prompt_job.rb
@@ -1,0 +1,13 @@
+module ScriptGenerator
+  module ForMotivation
+    class GenerateMotivationPromptJob < ApplicationJob
+      include JobErrorHandler
+
+      def perform(chat_id)
+        ProcessMotivationPrompts.call(chat_id:)
+      rescue StandardError => e
+        notify_script_generator_error(chat_id:, error: e)
+      end
+    end
+  end
+end
