+++ app/jobs/script_generator/for_motivation/generate_motivation_script_job.rb
@@ -1,0 +1,13 @@
+module ScriptGenerator
+  module ForMotivation
+    class GenerateMotivationScriptJob < ApplicationJob
+      include JobErrorHandler
+
+      def perform(chat_id, language = "en")
+        ProcessMotivationScript.call(chat_id:, language:)
+      rescue StandardError => e
+        notify_script_generator_error(chat_id:, error: e)
+      end
+    end
+  end
+end
