--- app/jobs/script_generator/generate_motivation_script_job.rb
@@ -1,11 +1,0 @@
-module ScriptGenerator
-  class GenerateMotivationScriptJob < ApplicationJob
-    include JobErrorHandler
-
-    def perform(chat_id, language = "en")
-      ScriptGenerator::ProcessMotivationScript.call(chat_id:, language:)
-    rescue StandardError => e
-      notify_script_generator_error(chat_id:, error: e)
-    end
-  end
-end
